requirements.txt
aniso8601==8.0.0
astroid==2.4.2
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-JWT==0.3.2
Flask-RESTful==0.3.8
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.0
pandas==1.0.5
pdfkit==0.6.1
PyJWT==1.4.2
pylint==2.5.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1
wkhtmltopdf==0.2
wrapt==1.12.1
xlrd==1.2.0
uwsgi	
sendgrid==6.4.3	
psycopg2-binary==2.8.5
